requirements.txt
streamlit
pdf2image
Pillow
pandas
azure-ai-formrecognizer
google-cloud-vision
openai
python-dotenv
sqlalchemy
psycopg2-binary
